Add node_modules from devdep to NODE_PATH
When installing pc-nrfconnect-devdep in other projects, it does
sometimes not flatten/dedupe the dependencies in devdep. The
dependencies end up in "pc-nrfconnect-devdep/node_modules" instead of
the root node_modules directory. When running f.ex. webpack, it will
then not find the dependencies, as it only looks in the root
node_modules directory.

To fix this, we can add "pc-nrfconnect-devdep/node_modules" to the
NODE_PATH environment variable when spawning build/test/lint processes
from nrfconnect-scripts.js.