Switch from node-sass to dart-sass
node-sass has a natively compiled portion, which leads to recurring
problems, e.g. with new node versions.

dart-sass (whose npm module is just called sass) on the other hand is
compiled to pure javascript and also the current reference
implementation.

Using dart-sass showed no major performance disadvantages and the
created output only differs in little things which should not lead to
different behaviour.

Please note, that this switch needs a change in the webpack
configuration. For the apps this is already done in this commmit, for
core this needs to be done when this version of devdep is used:

```
diff --git a/webpack.config.js b/webpack.config.js
index 60abac7..e04b300 100644
--- a/webpack.config.js
+++ b/webpack.config.js
@@ -47,10 +47,16 @@ module.exports = {
             exclude: /node_modules/,
         }, {
             test: /\.scss|\.css$/,
-            loaders: [
+            use: [
                 MiniCssExtractPlugin.loader,
-                require.resolve('css-loader'),
-                require.resolve('sass-loader'),
+                'css-loader',
+                {
+                    loader: 'sass-loader',
+                    options: {
+                        // eslint-disable-next-line global-require
+                        implementation: require('sass'),
+                    },
+                },
             ],
         }, {
             test: /\.(png|gif|ttf|eot|svg|woff(2)?)(\?[a-z0-9]+)?$/,
```