Rename project to pc-nrfconnect-shared
Because this may require that other packages using this package have to
be slightly adapted, this is a breaking change.

But please note, that they only can and have to be adapted if they
actually depend on version 4 (or later) of this package. And it is not
necessary, that all packages have to be adapted right away:

E.g. the core/launcher could already use version 4.0 of this package
under the name pc-nrfconnect-shared, while an app may still refer to
version 3.5 of it as pc-nrfconnect-devdep.