Convert react and react-redux to peer dependencies
It is generally considered better practice that libraries declare these
as peer dependencies and when someone uses this library
(pc-nrfconnect-shared), then the dependency on react and react-redux
should be declared by that someone (e.g. the launcher in our case).

Having direct dependencies also caused problems when using `npm link` to
link a live version of pc-nrfconnect-shared to pc-nrfconnect-launcher
during development.

Along the way I also updated the dependencies.