Make list order changes easier to follow
When a device changes it position in the list (because it is favorited
or the name is changed) it may jump to a different position which is
confusing. So this adds an animation, so that the transition is visually
shown on the screen.

As a bonus it is also now animated when devices are added or removed,
making it easier to see which devices come, go, or stay.