Make visibility bar always distinguisable
In the visibility bar, the background of the “show side panel” area had
the same colour as the padding in the main view. This looked strange if
both side panel and log are hidden. Resolution: Change the background of
the “show side panel” area to the background colour of the side panel if
the side panel is hidden.