Fix handling of clicking on selected device name
When clicking on the name of the selected device it was able to rename
it. But it should be the same when clicking on any device name: There
is no special action, it is just the same as clicking on the surrounding
area, in the case of the selector button the device list should open.