Better alignment of port names
Port names were centered, which looked okay for long names but out of
place for shorter names.

If they are short enough, they are now left aligned with the other text
above (device name, serial number, PCA number). If the port names are so
long that they would run to the right out of the device area, then they
port names are moved to the left just enough to fit.

The same logic (usually left aligned with other text, maybe moved left
into the empty space) is also applied to the other information fields
(PCA number and potentially the device name if is has a nickname) but
in practice this does not happen yet.