Do not use void as a callback return value
Following https://www.typescriptlang.org/docs/handbook/declaration-files/do-s-and-don-ts.html#return-types-of-callbacks
usually switch from `any` to `void` as return value for callbacks.

In the case of `releaseCurrentDevice` switch to `Promise<void> | void`
and add an explanation that this callback may return a promise in which
case the execution will not be continued until the promise is fulfilled.