Adapt the inline inputs width to the current value
With this the input can grow longer if needed and we do not need to
specify a width for fields that can contain fractions.

"overflow-x: hidden" is set on the input to prevent it from growing
too much, e.g. wider than the space that is provided by the side panel.