Check stricter order of the imports during linting
This enforces a stricter check for the order of the imports and
consequently updates all files where the order was not correct. Mostly
this order is what we already do, but we did deviate a bit from it in a
lot of places.

TL;DR: First group, then sort alphabetically.

The main rule is to split the imports into four groups in this order:
1. Very rare: Side effect imports (like `import 'focus-visible'`)
2. Normal package imports (like `import React from 'react'`)
3. Relative imports (like `import logger from '../logging'`)
4. Style imports (like `import './about.scss'`)

These groups are all separated by newlines.

Within these groups the imports are ordered alphabetically by the "from"
(`prop-types` before `redux`) and the imported items themselves are also
ordered alphabetically (`arrayOf` before `node`).

There is one special case for the ordering of package names: The package
`react` and all other packages beginning with `react` will be at the
beginning of their group. This respects the current habit of putting the
React import at the top.

You can see how these rules play out in the files in this project, best
in a file with many imports such as `App.jsx`.

The rather obscure configuration is commented in `eslintrc.json`. Even
though the JSON file format normally does not allow comments, ESLint
explicitly allows them in their configuration. This means some editors
might complain about the comments, but I preferred that to converting
the configuration to plain JavaScript, since that would have been
another breaking change for all packages using this package.