Commit package-lock.json
`pc-nrfconnect-shared` is (of course) not an app but used as a library
in other projects. My understanding was, that in these cases
`package-lock.json` should not be committed. That stems from the fact,
that `package-lock.json` is ignored by other packages, that depend on
`pc-nrfconnect-shared`.

What I learned now: This is not an undisputed practice. Some advocate
for commiting `package-lock.json` also for library packages.

One reason for this, which we now also experienced several times: By not
commiting `package-lock.json`, the build is not reproducable anymore,
because retrieving and building the package again at a later time may
lead to different dependencies being used.

So in this commit I added a `package-lock.json` and change the azure
build to use it.