Remove unneeded prop type
The property `children` is by default already included in
React.Component, so one usually does not need to specify it.

You _can_ specify it though, if you want it to have a more specific or
different type than the default. Here it was actually made mandatory,
but that seems to be unintended, so I removed it.