Fix the layout bug in the toolchain manager
The CSS classname `nrf-card` used in the `Card` component clashed with a
likewise classname in the toolchain manager.

To fix the problem but also demontrate how we can prevent such problems
in the future, this commit switches in the `Card` component from
traditional CSS classnames to CSS modules.

CSS modules can from now on also be used in other places, the filename
of the CSS file just needs to include `.module.`.