Pin the prettier version to exact version
This is a practice [recommended in the prettier
documentation](https://prettier.io/docs/en/install.html). Not doing it
so far led to unwanted behaviour: When updating to any new version of
`pc-nrfconnect-shared` in an app or the launcher, the formatting might
change, even though it is not mentioned in the Changelog.

This could happen, because pc-nrfconnect-shared until this commit
only depended on version `^2.1.1` of prettier. If e.g. a version 2.2.0
of prettier was released, installing any new version of
pc-nrfconnect-shared led to also installing prettier 2.2.0.