Fix: Correct copyright check for Windows
Line endings could be either \r\n or \n. And this not only depends on
the platform of the developer machine but also on personal settings,
since git might be configured to do line ending conversion.

So, this is a two part solution:

When parsing texts, we are lenient by using the regexp \r?\n, which
accepts either \r\n or \n as a line ending.

When rewriting copyright headers, we look for the first line ending in
the file and use that for the header we write. Thus, we try to preserve
the kind of line ending that already exists in a file.