Fix: Wrongly frozen properties
Because we added the react components of the main panes to the store,
any properties on them were frozen by Immer. This causes a problem in
the old link monitor.

Since we do not really need the components (only the names of the
panes), this commit removes the components from the store.