Fix detection of deviating jlink versions
The string `jlink_dll` was outdated and it was necessary to also search
in the depencies, which `getModuleVersion` now does.

Different from before this code now calls `getModuleVersions` twice.
But I think the code is easier to understand if `logLibVersions` does
not also return the list of versions. And the performance penality is
small, because `getModuleVersions` is fast, the code is ran only once
while starting an app, and also asynchronously.