Ignore all entries from `.gitignore` in linting
This required changing the format of the ESLint config from JSON to JS.

This brings along other nice things, e.g. that comments are now
compatible with the file format.

But if the files was previously referenced as `eslintrc.json` (which is
the case in all app configurations) then that must be changed.