Fix logging with bundled apps
The bug: When bundling shared with an app, clicking on “open log file”
did not open the log but showed an error in the log viewer: “Unable to
open log file: Error: Could not find file at path: undefined”. Also, the
log file created on disk was always empty.

The cause: If the app bundles shared, then the module `logging/index.ts`
exists twice: The one bundled with the launcher and the one bundled with
the app. The launcher calls `logger.addFileTransport` on the module
bundled with the launcher, so the module scoped variable `logFilePath`
is set in that module. When the app then calls the `logger` to log
something, this is done on the module bundled with the app, so the
module scoped variable `logFilePath` is not set in that module.

The fix: Instead of calling `logger.addFileTransport` from the launcher
call it during initialisation from within the app.