Add SerialPort wrapper implementation
Should work as a wrapper to the serialport library, which also bumps
the serialport version to 10, from version 8 that was previously used
in most apps.

The idea is to use the main process in launcher to be the process that
directly interfaces with the serial ports, and then allow renderer
processes to interface with the serial port through the main process.

The implementation in shared, the SerialPort wrapper, is to export
simplify the interaction between renderer process and main process.

This is by making the wrapper hide away the communication between main
and renderer, and exposing a similar API as the SerialPort library
exposes.