Write code more consistent and clean up
Make channel strings and variable names more consistent. Handle error
scenarios by using async await when opening and closing the SerialPort,
and escalate the error in order for the renderer to handle the error.

Keep logging in order to make it consistent between apps, but make them
more compact and consistent.