Add overwrite support when opening port
When a renderer process wants to open a port, it may add overwrite in
order to overwrite the current settings used by the port.

In the case that a port has changed settings, it will emit an
`serialport:on-changed` event with the new options, in order to let
the other renderer processes to adjust it's UI accordingly.