Change explicit value prop for Slider component
Similar to the same change before for `NumberInlineInput`: Motivation
for this is that when specifying explicit values, the mandatory
properties for `min` and `max` where not needed anymore and might even
contradict the values.

Slight change for the behaviour when requesting ticks for the slider:
There is still an error message shown, that this is not implemented yet,
but contrary to before no ticks at all are shown in this case, instead
of wrong ticks as before.