Fix: Error when installing shared as dependency
This was sometimes triggered when installing shared as a dependency with
npm@7 or later.

The `prepare` script of `shared` is at least sometimes executed in that
case, which failed because it contained `husky install`. That in turn
expects a `.git` directory. So now the script first checks if that
folders exists (which is only correct if `shared` was cloned for
development) and only then executes `husky install`.