Fix: Reactivate the check for unused variables
The check is useful most of the times.

`@typescript-eslint/no-unused-vars` was turned off because
`no-unused-vars` was already turned off. But this is wrong:
`no-unused-vars` was turned off because it is not well suited for
TypeScript, by having false positives and false negatives for certain
TypeScript structures. So, it is superseded by
`@typescript-eslint/no-unused-vars`.

So, this change turns `@typescript-eslint/no-unused-vars` back on.

This also removes the configuration to turn off `no-unused-vars` because
that is already done by `plugin:@typescript-eslint/recommended`.