Switch to using nrf-fw-update-bundle-lib for dfu generation
This is an initial attempt at using the library instead of the current
typescript-based dfu generation code.

These changes aim for minimal disruption and I haven't removed any
dead code that is no longer used.

This change will break the pc-nrfconnect-programmer code base which
relies on the exported createDfuZipBuffer function which we've changed
here.