Use more flexible 'input' function
Instead of using if-statements to get down to the correct function
call to the library.

We let the library take care of it instead. It makes the sdId and sdReq
stuff a bit more confusing and I'm not sure if I have it right but we
can try to sort that out over time.