Small improvements to Conflict Dialog
Prettifies values of settings to be displayed, naively using JSON.stringify
was not working as we wanted. E.g. it was showing all strings with
quotes, which we prefer to avoid.

Gives a red background to the values that are conflicting, so that the
user may quickly identify the difference. This was also sufficient in
order to drop a separate `conflicting` box of settings.

More precise naming. `Active` refers to the settings that are in use,
and `Selected` are the settings that are attempted to be used by the
current application.