Fix error when called in launcher
The launcher calls the default export from `sandbox.ts` and that in turn
expected that the `package.json` of the app to use was defined during
compilation. But we do that only while compiling apps (because only they
have a package.json with all the fields we expect). And when the
launcher calls `sandbox.ts` the `package.json` is not really needed,
because the launcher explicitly states which version of the module to
install.

So this changes the logic slightly that if the version of the module to
install is already specified, `sandbox.ts` doesn't look at
`package.json` any longer (and thus also doesn't fail if it isn't
defined).