Allow parent component access to validity in NumberInlineInput
In order to give parent component of the NumberInlineInput access to the
validity of the field, expose a callback function that can be used to
access the validity.

In this iteration, if the valueIsValidCallback is defined, then we do
not propagate the validity to the InlineInput component, by always
setting the `isValid` property to `true`.