Fix: Uncaught exception when invoking nrfutil
When spawning the nrfutil process fails in certain ways, an uncaught
exception in the main process got thrown.

The “certain ways” make this a bit hard to reproduce: On macOS this
happened, when the nrfutil executable did not have the executable file
mode. Usually this should not happen, because we set that mode ourselves
correctly.

This resulted in an ugly dialog, outside of the launcher getting
displayed by electron.

To circumvent this, adding a handler for the `error` event is
sufficient. With this, the exception is not uncaught any longer.