Restrict type of nrfModules to a single semver
In theory the package.json of apps could contain something like

```json
    "nrfConnectForDesktop": {
        "nrfutil": {
            "device": [
                "2.0.2",
                "2.1.2",
            ]
        }
    }
```

But in practice we ignore all but the first version. To not mislead
people into believing they can specify multiple versions and it will be
effective, this is now restricted to a single version. If more versions
are specifiec, this will now cause an error while parsing.

If, in the future, we decide to support multiple versions, this can be
changed again easily.