Correct links
Corrected the links for nRF9161 DK, Thingy:91 X, nRF9151 DK, and
nRF54L15 DK.

For nRF54H20 DK there is no product page yet, but if it gets created at
some point, then the corrected links are the ones we expect.

There currently is no nRF9131 EK, so the links for that are removed.