Make open* functions to be IPC calls
openFile, openFileLocation, and openUrl were invoking code in the main
process and are now refactored to use IPC calls. This became necessary
at least for openFileLocation because it might hang when called on macOS
from the renderer process. But it means that the handlers for the IPC
now need to be registered by the launcher, which will be done from 5.2.0
onwards, so apps using this version of shared must depend at least on
that launcher version.