Specify core version of nrfutil
Apps must now specify what core version of nrfutil they depend on. This
is to ensure that more of the environment the app runs in is locked for
an app and future updates, in nrfutil do not break the app unexpectedly.

If an app e.g. specifies that it requires nrfutil core 8.0.0 and nrfutil
device 2.9.0, then that sandbox will be created in the path
`nrfutil-sandboxes/8.0.0/device/2.9.0`.

If the launcher encounters an old app, which does not specify the core
version yet, it will default to using nrfutil core 8.0.0, but use the
old path, because those apps expect it there, e.g.
`nrfutil-sandboxes/device/2.9.0`.

(On macOS with Apple Silicon there would be an additional segment
`arm64/` after `nrfutil-sandboxes/` in those examples)