Fix error when verbose logging is turned on
The error was introduced in #1003 when `this.log` was passed to the new
CollectingResultParser. But because `log` was a regular method, `this`
in it pointed to the CollectingResultParser instance, not the
NrfutilSandbox instance. So turning `log` into an arrow function fixed
the issue.

The error was only triggered when verbose logging was turned on, because
only in this case log events were emitted.