Fix: No problematic nRF Util core version check
The check was unnecessary, because it is only called when a sandbox is
created, so there is no nRF Util core yet. And it was problematic,
because it led to the latest version (or more concret to the one in
https://files.nordicsemi.com/artifactory/swtools/external/nrfutil/index/bootstrap.json)
of the core being installed initially and that version might not run
correctly, which is why we introduced the ability to specify a core
version in the first place.