Remove unpublish workflow
With the new, trusted publisher based workflow, it is no longer possible
to unpublish packages without an npm token.

But apart from testing purposes, we never needed this anyhow, so there
is little pain in removing it.

If in the future a version needs to be unpublished, someone who has the
appropriate access rights at the npm org @nordicsemiconductor has to do
it manually.