Add new npm publish switch workflow
With the trusted publisher model, npm now allows only a single workflow
to be registered as the trusted publisher. But this workflow can then
invoke other workflows, which may do the actual publishing.
https://docs.npmjs.com/trusted-publishers

So the new `npm-publish-switch.yml` workflow will act as a switch which:

- If run on main branch as a result of a push, will check try to
  automatically release a new version
- Can be run directly to explicitly publish a new version to npm (which
  only makes sense when running it on a branch apart from main, because
  there the auto-publish workflow will take care of it)

The idea of this setup is also shown in
https://github.com/orgs/community/discussions/174507#discussioncomment-14723818

Because trusted publishing requires npm CLI version 11.5.1 or later, 
this currently also means, that we have to update npm before running 
`npm publish`.

Additionally `npm-publish-switch.yml` also needs to be registered as the 
trusted publisher in the npm settings for the pc-nrfconnect-shared 
package.