Fix: Failing app release still creates GH release
The nordic-publish script checks that the app is in a releasable state,
e.g. if an official release is done that there is an appropriate entry
in the changelog.

But the GitHub release was created before that check.

So, if the the check failed, a GitHub release was still created. So that
had to be removed manually before trying another release.

Because of this, the nordic-publish is now run before creating the
GitHub release.

Even better would be to run _all_ release checks before either creating
the GitHub release or releasing to Artifactory but that would be more
effort than this simpler fix.